requirements.txt
# =============================================================================
# SongFinder Requirements
# =============================================================================
# These versions are tested and known to work together.
# Run: pip install -r requirements.txt
# =============================================================================

# Core ML Framework (install these first via setup_env.sh)
torch==2.10.0
torchvision==0.25.0
torchaudio==2.10.0
torch-geometric==2.7.0
torchmetrics==1.8.2

# Audio Processing
librosa==0.11.0
soundfile==0.13.1
mutagen==1.47.0
audioread==3.1.0

# Scientific Computing (IMPORTANT: numpy<2.0 for scipy/librosa compatibility)
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.8.0

# Utilities
tqdm==4.67.1
PyYAML==6.0.3
matplotlib==3.10.8
networkx==3.6.1
timm==1.0.24
prettytable==3.17.0

# Note: faiss-gpu is installed via conda in setup_env.sh (not pip)
